isAdmin-Errorhandle

+++ src/routes/admintarif.js
@@ -1,5 +1,6 @@
 const express = require('express')
 const router = express.Router()
+const {isAdmin} = require('../lib/auth');
 
 const pool = require('../database')
 
@@ -8,7 +9,7 @@ router.get('/ajoutTarif', (req, res)=> {
 })
 
 
-router.post('/ajoutTarif', async (req, res)=> {
+router.post('/ajoutTarif',isAdmin, async (req, res)=> {
   const {PrixTarif,SuplementNbPersSup4 } = req.body;
   const newTarif = {PrixTarif,SuplementNbPersSup4 };
   
@@ -18,7 +19,7 @@ router.post('/ajoutTarif', async (req, res)=> {
 })
 
 /*Info pour le form de modification semaint */ 
-router.get('/modifTarif/:NumTarif',async(req,res) => {
+router.get('/modifTarif/:NumTarif',isAdmin,async(req,res) => {
   const {NumTarif}=req.params;
   const tarifs = await pool.query('SELECT * FROM tarif WHERE NumTarif = ?', [NumTarif]);
   
@@ -26,7 +27,7 @@ router.get('/modifTarif/:NumTarif',async(req,res) => {
 
 })
 /*Modification de Tarif via form*/
-router.post('/modifTarif/:NumTarif', async(req,res)=>{
+router.post('/modifTarif/:NumTarif',isAdmin, async(req,res)=>{
   const {NumTarif}=req.params;
   console.log(NumTarif)
   const {PrixTarif,SuplementNbPersSup4} = req.body;
@@ -34,7 +35,7 @@ router.post('/modifTarif/:NumTarif', async(req,res)=>{
   const newTarif = { PrixTarif,SuplementNbPersSup4};
   
  
- await pool.query('UPDATE tarif  set ? WHERE NumTarif = ?', [newTarif,NumTarif]);
+ await pool.query('UPDATE tarif  set ? WHERE NumTarif = ?',isAdmin, [newTarif,NumTarif]);
  req.flash('success','Tarif n°'+ NumTarif+' modifiée');
  res.redirect('/admin/listeTarifs');
 
@@ -44,7 +45,7 @@ router.post('/modifTarif/:NumTarif', async(req,res)=>{
 
 router.get('/supprTarif/:NumTarif',async (req,res)=>{
   const {NumTarif} = req.params;
-  await pool.query ('DELETE FROM tarif WHERE NumTarif=?',[NumTarif]);
+  await pool.query ('DELETE FROM tarif WHERE NumTarif=?',isAdmin,[NumTarif]);
   req.flash('message','Tarif n°' +NumTarif+ ' supprimée');
   res.redirect('/admin/listeTarifs');
